# 05_openapi_final.yaml
openapi: 3.1.0
info:
  title: Kaveri Stays API
  version: "1.0.0"
  summary: Reveal specification for the Kaveri Stays API assignment.
  description: |
    # Read this before you read the paths

    This is ONE reasonable design. It is not the only one. Several of the choices
    below are genuinely arguable and the Interlude expects you to argue with them.
    A few are not arguable, and those are marked.

    ## Conventions this specification commits to

    **Dates.** Every date is ISO 8601 `YYYY-MM-DD`. A stay is the half-open interval
    `[check_in, check_out)`. The check-out date is NOT part of the stay. This is what
    makes same-day turnover work (rule 5), and it is NOT arguable.

    **Money.** Every amount is a decimal string, never a JSON number. `"4500.00"`,
    not `4500.0`. Floats lose money.

    **Empty collections.** A list endpoint with no matching rows returns 200 and an
    empty `items` array. It never returns 404. 404 means the container does not exist,
    not that it is empty.

    **401 versus 403.** 401 means the server does not know who you are: no token, an
    expired token, a forged token. 403 means it knows exactly who you are and you may
    not do this. This distinction is NOT arguable.

    **Request bodies.** Every request schema sets `additionalProperties: false`. A
    field you did not declare is rejected rather than ignored. This is what defeats
    attacks 8.2 and 8.8 structurally rather than by remembering to check.

...

...

...

...

...

  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Register a guest account.
      description: |
        Creates a guest account and nothing else. There is no path through this
        endpoint to any other role. `additionalProperties: false` is what makes
        attack 8.2 fail; do not rely on remembering to strip a `role` field.
      security: []
      x-roles: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Account created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Exchange credentials for a token pair.
      description: |
        The failure response is identical for an unknown email and a wrong password.
        Anything else answers attack 8.14 for the attacker.
      security: []
      x-roles: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refresh
      summary: Rotate a refresh token.
      description: |
        Refresh tokens are single use. Presenting one that has already been rotated
        is 401, and a correct implementation also invalidates the whole token family,
        because reuse means the token leaked. This is attack 8.6.
      security: []
      x-roles: [public]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: New token pair issued; the presented refresh token is now dead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke the caller's refresh token.
      x-roles: [guest, staff, manager, owner]
      responses:
        '204':
          description: Revoked. Nothing to return.
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me:
    get:
      tags: [auth]
      operationId: getMe
      summary: The caller's own record.
      x-roles: [guest, staff, manager, owner]
      responses:
        '200':
          description: The caller.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token in `Authorization: Bearer <token>`. Short lived. Obtain a new one
        from `/auth/refresh` rather than by logging in again.
...
  responses:
    Unauthorized:
      description: No token, an expired token, or a token that does not verify.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: { code: unauthenticated, message: Authentication required. }
...
    Conflict:
      description: The request conflicts with the current state of the system.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Unprocessable:
      description: |
        The request parsed and a field is unacceptable. Includes guest counts above
        the room type maximum, check-out on or before check-in, ratings outside 1 to 5,
        and any field not declared in the request schema.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: validation_failed
              message: Request validation failed.
              detail:
                - field: guests
                  reason: Exceeds maximum occupancy of 2 for room type Standard.
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
      headers:
        Retry-After:
          description: Seconds until the caller may retry.
          schema: { type: integer }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      description: |
        Every non-2xx response in this API uses this shape. One handler produces it.
        If you are constructing this dict in more than one module, consolidate.
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable, snake_case. Clients branch on this.
              examples: [room_unavailable, illegal_transition, validation_failed]
            message:
              type: string
              description: Human-readable. Safe to show a guest. Never a raw Postgres message.
            detail:
              description: Optional structured context. Never leaks table or column names.
            request_id:
              type: string
              description: Correlates this response with your logs.
...
    Role:
      type: string
      enum: [guest, staff, manager, owner]
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, full_name]
      description: |
        No `role` field exists here and none is accepted. Sending one is 422, which is
        attack 8.2 refused structurally.
      properties:
        email:     { type: string, format: email }
        password:  { type: string, minLength: 10, writeOnly: true }
        full_name: { type: string, minLength: 1, maxLength: 120 }
        phone:     { type: string }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:    { type: string, format: email }
        password: { type: string, writeOnly: true }
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token: { type: string }
    TokenPair:
      type: object
      additionalProperties: false
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:  { type: string }
        refresh_token: { type: string }
        token_type:    { type: string, const: bearer }
        expires_in:    { type: integer, description: Access token lifetime in seconds. }
    Me:
      type: object
      additionalProperties: false
      required: [id, email, full_name, role]
      description: No password hash appears in this schema or any other. Task 4.6.
      properties:
        id:          { type: integer }
        email:       { type: string, format: email }
        full_name:   { type: string }
        role:        { $ref: '#/components/schemas/Role' }
        property_id:
          type: [integer, "null"]
          description: The property this staff member or manager belongs to. Null for guest and owner.
...